pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voice-generation-agent"
version = "1.0.0"
description = "CLI agent for generating voice from documents and scripts"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = ["voice", "tts", "text-to-speech", "cli"]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "google-cloud-texttospeech>=2.16.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "python-docx>=1.1.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
voice-agent = "src.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]
